[FLINK-24270][checkpoint] Refactor the tests related to the VertexFinishedStateChecker.
